pyproject.toml
[project]
name = "zoialib"
version = "0.2.0"
description = "A patch librarian for the Empress ZOIA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rich>=14.1.0",
    "tomlkit>=0.13.3",
    "typer-slim>=0.19.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
zoialib = "zoialib.__main__:app"
